+++ packages/api/src/feature/club/club.module.ts
@@ -9,13 +9,13 @@ import UserModule from "../user/user.module";
 import { ClubController } from "./controller/club.controller";
 import { ClubDelegateDRepository } from "./delegate/club.club-delegate-d.repository";
 import { DelegateModule } from "./delegate/delegate.module";
-import { ClubRoomTRepository } from "./repository/club.club-room-t.repository";
-import ClubStudentTRepository from "./repository/club.club-student-t.repository";
-import ClubTRepository from "./repository/club.club-t.repository";
-import { DivisionPermanentClubDRepository } from "./repository/club.division-permanent-club-d.repository";
-import { ClubGetStudentClubBrief } from "./repository/club.get-student-club-brief";
-import { ClubPutStudentClubBrief } from "./repository/club.put-student-club-brief";
-import ClubRepository from "./repository/club-old.repository";
+import { ClubRoomTRepository } from "./repository-old/club.club-room-t.repository";
+import ClubStudentTRepository from "./repository-old/club.club-student-t.repository";
+import ClubTRepository from "./repository-old/club.club-t.repository";
+import { DivisionPermanentClubDRepository } from "./repository-old/club.division-permanent-club-d.repository";
+import { ClubGetStudentClubBrief } from "./repository-old/club.get-student-club-brief";
+import { ClubPutStudentClubBrief } from "./repository-old/club.put-student-club-brief";
+import ClubRepository from "./repository-old/club-old.repository";
 import ClubPublicService from "./service/club.public.service";
 import { ClubService } from "./service/club.service";
