+++ packages/api/src/feature/club/repository-old/club.get-student-club-brief.ts
