+++ packages/api/src/feature/common-space/common-space.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 
 import { DrizzleModule } from "@sparcs-clubs/api/drizzle/drizzle.module";
 import ClubModule from "@sparcs-clubs/api/feature/club/club.module";
-import ClubStudentTRepository from "@sparcs-clubs/api/feature/club/repository/club.club-student-t.repository";
+import ClubStudentTRepository from "@sparcs-clubs/api/feature/club/repository-old/club.club-student-t.repository";
 import { SemesterModule } from "@sparcs-clubs/api/feature/semester/semester.module";
 import UserModule from "@sparcs-clubs/api/feature/user/user.module";
